Adds tilde expansion to `fs.monitor`
Also fixes some clippy warnings.
Fixes: #518
