+++ src/tools/mod.rs
@@ -1,4 +1,5 @@
 mod crash_handler;
 pub mod patch;
+pub mod util;
 
 pub use self::crash_handler::register_panic_hook;
